feat: Add FileView component for displaying files
The FileView component is added to display a list of files. It takes an array of files as a prop and renders FileViewTile components for each file. If there are no files, a message is displayed indicating that there are no files. This component is used in the MainPage component to show the list of files.